@@ -36,54 +40,39 @@
     "organization": false
   },
   "devDependencies": {
-    "@aws-cdk/assert": "^1.111.0",
     "@types/jest": "^26.0.24",
-    "@types/node": "^10.17.0",
-    "@typescript-eslint/eslint-plugin": "^4.29.2",
-    "@typescript-eslint/parser": "^4.29.2",
-    "esbuild": "^0.12.20",
-    "eslint": "^7.32.0",
+    "@types/node": "^14",
+    "@typescript-eslint/eslint-plugin": "^5",
+    "@typescript-eslint/parser": "^5",
+    "aws-cdk-lib": "^2.27.0",
+    "constructs": "^10.0.5",
+    "esbuild": "^0.14.43",
+    "eslint": "^8",
     "eslint-import-resolver-node": "^0.3.6",
-    "eslint-import-resolver-typescript": "^2.4.0",
-    "eslint-plugin-import": "^2.24.0",
-    "jest": "^26.6.3",
-    "jest-junit": "^12",
-    "jsii": "^1.33.0",
-    "jsii-diff": "^1.33.0",
+    "eslint-import-resolver-typescript": "^2.7.1",
+    "eslint-plugin-import": "^2.26.0",
+    "jest": "^27",
+    "jest-junit": "^13",
+    "jsii": "^1.60.1",
+    "jsii-diff": "^1.60.1",
     "jsii-docgen": "^1.8.110",
-    "jsii-pacmak": "^1.33.0",
-    "json-schema": "^0.3.0",
-    "npm-check-updates": "^11",
-    "projen": "^0.27.30",
-    "source-map-support": "^0.5.19",
+    "jsii-pacmak": "^1.60.1",
+    "json-schema": "^0.4.0",
+    "npm-check-updates": "^12",
+    "projen": "^0.58.4",
+    "source-map-support": "^0.5.21",
     "standard-version": "^9",
-    "ts-jest": "^26.5.6",
-    "typescript": "^4.3.5"
+    "ts-jest": "^27",
+    "typescript": "^4.7.3"
   },
   "peerDependencies": {
-    "@aws-cdk/aws-events": "^1.111.0",
-    "@aws-cdk/aws-events-targets": "^1.111.0",
-    "@aws-cdk/aws-iam": "^1.111.0",
-    "@aws-cdk/aws-lambda": "^1.111.0",
-    "@aws-cdk/aws-lambda-nodejs": "^1.111.0",
-    "@aws-cdk/aws-logs": "^1.111.0",
-    "@aws-cdk/aws-stepfunctions": "^1.111.0",
-    "@aws-cdk/aws-stepfunctions-tasks": "^1.111.0",
-    "@aws-cdk/core": "^1.111.0",
-    "constructs": "^3.2.27"
+    "aws-cdk-lib": "^2.27.0",
+    "constructs": "^10.0.5"
   },
   "dependencies": {
-    "@aws-cdk/aws-events": "^1.111.0",
-    "@aws-cdk/aws-events-targets": "^1.111.0",
-    "@aws-cdk/aws-iam": "^1.111.0",
-    "@aws-cdk/aws-lambda": "^1.111.0",
-    "@aws-cdk/aws-lambda-nodejs": "^1.111.0",
-    "@aws-cdk/aws-logs": "^1.111.0",
-    "@aws-cdk/aws-stepfunctions": "^1.111.0",
-    "@aws-cdk/aws-stepfunctions-tasks": "^1.111.0",
-    "@aws-cdk/core": "^1.111.0"
+    "aws-cdk-lib": "^2.27.0",
+    "constructs": "^10.0.5"
   },
-  "bundledDependencies": [],
   "keywords": [
     "aws-cloudwatch-events",
     "aws-lambda",
